Revert "feat: Implement note editor metadata drawer front - EXO-71928_EXO-719…"
This reverts commit 4ba2578cfff5febe16523ecaaa589e3a9d70307d.
