+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_en.properties
@@ -50,12 +50,6 @@ notes.edit.editNotes=Edit
 notes.save.success.message=Note saved successfully
 notes.view.label=View
 
-notes.metadata.properties.label=Properties
-notes.metadata.featuredImage.label=Featured images
-notes.metadata.featuredImage.add.label=Add an image
-notes.metadata.summary.label=Summary
-notes.metadata.add.summary.placeholder=Write a text to summarize the content
-notes.metadata.featuredImage.edit.label=Edit The Featured Image
 
 popup.confirm=Confirm
 popup.msg.confirmation=Confirmation
