+++ notes-webapp/src/main/webapp/WEB-INF/conf/configuration.xml
@@ -32,6 +32,5 @@
   <import>war:/conf/wiki/search-configuration.xml</import>
   <import>war:/conf/wiki/ckeditor-configuration.xml</import>
   <import>war:/conf/wiki/metadata-configuration.xml</import>
-  <import>war:/conf/wiki/feature-flags-configuration.xml</import>
 
 </configuration>
