--- notes-webapp/src/main/webapp/WEB-INF/conf/wiki/feature-flags-configuration.xml
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
- * This file is part of the Meeds project (https://meeds.io/).
- *
- * Copyright (C) 2024 Meeds Association contact@meeds.io
- *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 3 of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
--->
-<configuration xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-               xsi:schemaLocation="http://www.exoplatform.org/xml/ns/kernel_1_3.xsd http://www.exoplatform.org/xml/ns/kernel_1_3.xsd"
-               xmlns="http://www.exoplatform.org/xml/ns/kernel_1_3.xsd">
-  <component>
-    <key>NotesEditorMetadataDrawerFeatureProperties</key>
-    <type>org.exoplatform.container.ExtendedPropertyConfigurator</type>
-    <init-params>
-      <properties-param>
-        <name>NotesEditorMetadataDrawerFeatureProperties</name>
-        <description>Metadata drawer Feature enablement flag</description>
-        <property name="exo.feature.notesEditorMetadataDrawer.enabled"
-                  value="${exo.feature.notesEditorMetadataDrawer.enabled:false}"/>
-      </properties-param>
-    </init-params>
-  </component>
-
-  <external-component-plugins>
-  <target-component>org.exoplatform.groovyscript.text.TemplateService</target-component>
-  <component-plugin>
-    <name>UIPortalApplication-head</name>
-    <set-method>addTemplateExtension</set-method>
-    <type>org.exoplatform.groovyscript.text.TemplateExtensionPlugin</type>
-    <init-params>
-      <values-param>
-        <name>templates</name>
-        <description>The list of templates to include in HTML Page Header with UIPortalApplication.gtmpl</description>
-        <value>war:/groovy/webui/workspace/UINotesHeadTemplate.gtmpl</value>
-      </values-param>
-    </init-params>
-  </component-plugin>
-  </external-component-plugins>
-</configuration>
