+++ notes-webapp/src/main/webapp/groovy/webui/workspace/UINotesHeadTemplate.gtmpl
@@ -1,10 +1,11 @@
 <%
   import org.exoplatform.commons.api.settings.ExoFeatureService;
+  import org.exoplatform.portal.application.PortalRequestContext;
   def rcontext = _ctx.getRequestContext();
   ExoFeatureService featureService = uicomponent.getApplicationComponent(ExoFeatureService.class);
   def userName = rcontext.getRemoteUser();
 %>
 
 <script type="text/javascript" id="notesHeadScripts">
-    eXo.env.portal.notesEditorMetadataDrawerEnabled = <%=featureService.isFeatureActiveForUser("notesEditorMetadataDrawer", userName)%>;
+    eXo.env.portal.notesMultilingual = <%=featureService.isFeatureActiveForUser("NotesMultilingual", userName)%>;
 </script>
