+++ notes-webapp/src/main/webapp/skin/less/notes/notes.less
@@ -124,11 +124,6 @@
         height: 21px;
       }
 
-      .metadata-button-icon {
-        width: 20px;
-        height: 21px;
-      }
-
       .notesActions {
         .notesFormButtons {
           .notesFormLeftActions {
@@ -210,18 +205,6 @@
   .notes-application-content {
     overflow: auto;
     min-height: calc(~"100vh - 350px");
-  }
-}
-
-#editorMetadataDrawer {
-  .add-image-area {
-    background: @primaryBackground;
-  }
-
-  .cke_bottom {
-    height: auto;
-    opacity: 0.4;
-    background: transparent;
   }
 }
